Update translation functions; weapons cleanup

+++ Common/Game/animation.h
@@ -74,7 +74,7 @@ void AnimateItem(ItemInfo* item);
 bool HasStateDispatch(ItemInfo* item, int targetState = -1);
 bool TestLastFrame(ItemInfo* item, int animNumber = -1);
 
-void TranslateItem(ItemInfo* item, short orient, int forward, int vertical = 0, int lateral = 0);
+void TranslateItem(ItemInfo* item, short orient, float forward, float vertical = 0.0f, float lateral = 0.0f);
 void TranslateItem(ItemInfo* item, Vector3Shrt orient, float distance);
 void SetAnimation(ItemInfo* item, int animIndex, int frameToStart = 0);
